pyproject.toml
[tool.poetry]
name = "tableau-mcp"
version = "0.1.0"
description = ""
authors = ["samiribrh <samiribrahimov2277@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.2"
tableau = "^0.0.3"
tableau-api-lib = "^0.1.50"
tableauserverclient = "^0.38"
mcp-server = "^0.1.4"
mcp = "^1.21.0"
pantab = "^5.2.2"
pandas = "^2.3.3"
python-dotenv = "^1.2.1"
openpyxl = "^3.1.5"
fastapi = "^0.121.3"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
ollama = "^0.6.1"
python-multipart = "^0.0.20"
pydantic = "^2.12.4"
pydantic-settings = "^2.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
